feat: SoftDeletableEntity 추가
