+++ src/main/java/homeTry/common/entity/SoftDeletableEntity.java
@@ -1,0 +1,18 @@
+package homeTry.common.entity;
+
+import jakarta.persistence.MappedSuperclass;
+
+@MappedSuperclass
+public abstract class SoftDeletableEntity extends BaseEntity {
+
+    private boolean isDeprecated;
+
+    public boolean isDeprecated() {
+        return isDeprecated;
+    }
+
+    public void markAsDeprecated() {
+        this.isDeprecated = true;
+    }
+
+}
